core: introduce tiny tuples
Tuple bsize was stored in 4 byte uint32_t field. Tuple data_offset was
stored in 2 byte uint16_t field. Now the tuple contains flexible array
at the end (0 or 4 bytes), allowing to put it in the same way.
On the other hand, if it is possible, bsize and data_offset use 1 byte
each. Such tuples are called tiny tuples. They only require 6 bytes
instead of 10.
As long as struct tuple is reassembled to achieve it and is now of
variable size, inherited struct vy_stmt is also reworked now. It's
members (except struct tuple itself) are not directly listed anymore.
They were not used directly before and now they are not any more
available to avoid accidental mistakes. lsn, type and flags of
struct vy_stmt have to be accessed only using getters and setters now.

Part of #5385