iproto: fix comment and add assert on destruction
The comment in tx_process_destroy() about obuf destruction was
wrong. Memory for them is actually allocated from tx-belonging
slab cache and tx_process_destroy() obviously happens in tx, so
the comment is fixed to reflect the reality.
It is also implied that connection is in
IPROTO_CONNECTION_DESTROYED state in tx_process_destroy(). Now it
is verified with assert().

Part of #3776