tpch: use normal perf image instead of patched one
TPCH benchmark used to be launched with numerous patches applied
to tarantool's source code. Not only it does not bench actual
tarantool, it also adds extra complications in benchmarking process.
Recently we got pipeline failed only because of tpch benchmark as it
would fail to launch because it failed to apply all of the needed
patches. Instead of fixing the patches we'd better get rid of that
patching in the first place.