test: make strings compatible with Python 3
The largest change in Python 3 is the handling of strings.
In Python 2, the str type was used for two different
kinds of values - text and bytes, whereas in Python 3,
these are separate and incompatible types.
Patch converts strings to byte strings where it is required
to make tests compatible with Python 3.

Part of #5538

(cherry picked from commit 5c24c5ae90cb88e8657e4330100dd53372021a46)