test: revise zstd symbols list in static-build
In the scope of #5095 we've added a test that checks availability of
explicitly listed symbols. But the list wasn't generated properly and
included two private functions.

Nowadays within #5716 during zstd update we've encountered that the
`ZSTD_malloc` was renamed into `ZSTD_customMalloc` and the test broke.

This patch fixes the static-build test.