tuple: introduce unit perf test based on gbench
Test consists of two parts: the first one checks the performance of access
to indexed tuple fields, the second one checks the performance of memory
allocation and deallocation operations for tuples for typical workload.
Test is available only in release mode and in case when google benchmark
installed in system, otherwise it's silently ignored.

Part of #5385