test: update test-run (bump tarantool-python)
Update the test-run submodule with updating its tarantool-python
submodule with the Response.__str__() change, which unifies behaviour
between Python 2 and Python 3. We lean on this in the
box-py/call.test.py test. See more in [1].

Updated the box-py/iproto.test.py test, because it is confused by the
new IPROTO_SQL_INFO_* constants, introduced in the tarantool-python
connector in [2].

Part of https://github.com/tarantool/test-run/issues/20
Part of #5652

[1]: https://github.com/tarantool/test-run/pull/264
[2]: https://github.com/tarantool/tarantool-python/commit/9d3573e36199ee5ebc360f9017190ada9726bb23