github-ci: add link to log in Telegram message
The link to the failed job log is not saved in other GitHub Actions
links. At the same time this link can be obtained via the following
command using GitHub API:

 $ curl -s https://api.github.com/repos/tarantool/tarantool/actions/runs/${{  github.run_id }}/jobs | jq -r '.jobs[0].html_url'

The obtained link for the particular run has the following format:
https://github.com/tarantool/tarantool/runs/<id>, where 'id' is the
number of the job.

This link is added to the Telegram message for the failed job which ID
is set as suggested in GitHub docs[1].

[1]: https://docs.github.com/en/actions/reference/workflow-commands-for-github-actions#example-6

(cherry picked from commit 37d53e2a81b3e73445c3ebc4cdb27d16c78811c9)