wal: change WAL_ROWS_PER_YIELD to a power of 2
row_count % WAL_ROWS_PER_YIELD == 0 is tested by both local recovery and relay
on each recovered row, so it makes sense to have this constant a power of two,
so that the compiler could replace '%' with bitwise and.

WAL_ROWS_PER_YIELD used to be 32000, so change it to the closest power
of two: 32768 == 2 ^ 15.

Follow-up #5762

(cherry picked from commit 4842c090f87115792b3c4397e1ea020a4efe51c6)