sql: introduce mem_str()
This patch introduces mem_str() which allows to receive value of MEM as
a string. Due to the limitations of static_alloc(), this function cannot
be used to safely retrieve a value of MEM converted to string. This
function is suitable for debugging, displaying the value in an error
message, etc.

Part of #5818