github-ci: ubuntu-20.04 misses createrepo package
Found that ubuntu-latest that switched in Github Actions from
ubuntu-18.04 to ubuntu-20.04, doesn't have createrepo package
available in Ubuntu repositories to be installed. It happened
because createrepo written on python2, which is not supported
by ubuntu-20.04. Anyway createrepo tool was rewritten on C to
make it available for the later distributions, its new naming
is createrepo_c. It is already avaliable in Ubuntu 21.04, but
still not backported to 20.04 version. So we need to wait its
avalaibility in Ubuntu 22.04 either its backported version in
20.04 [1]. As for now we need to use ubuntu-18.04 in Actions.

[1]: https://answers.launchpad.net/createrepo/+question/690448