test: don't use IBUF_SHARED in msgpack test
It is used only to check that 'struct ibuf *' can be passed to
encode() functions. But soon IBUF_SHARED will be deleted, and its
alternative won't be yield-tolerant. This means it can't be used
in this test. There are yields between the buffer usages.

It is also used in the other tests, but there IBUF_SHARED is used
without yields in between, so its replacement will be easy.

Part of #5632