sql: introduce mem_get_int()
This patch introduces mem_get_int() function. This function is used to
receive integer value from MEM. If value of MEM is not integer, it is
converted to integer if possible. MEM is not changed.

Part of #5818

(cherry picked from commit dbdc4eec848e16c83eb4a7150935cde39bddcb2c)