sql: introduce mem_get_uint()
This patch introduces mem_get_uint() function. This function is used to
receive unsigned value from MEM. If value of MEM is not unsigned, it is
converted to unsigned if possible. MEM is not changed.

Part of #5818

(cherry picked from commit bde6d26e968143010e20448dcb8fb0e53f9adfb0)