lib: fix memory leak in rope_insert
Found that in case of exiting the rope_insert function with an error
some nodes are created but not deleted.

This commit fixes it and adds the test.

Test checks  that in case of this error the number of
allocated nodes and the number of freed nodes are the same.

Closes #5788

(cherry picked from commit 519408009eb57b2e7674e5be8e5f2241a9e306c8)