swim: add SO_BROADCAST option
Swim node couldn't talk to broadcast network interfaces because
the option SO_BROADCAST wasn't set.

It worked fine for localhost broadcast, but failed for all the
other IPs. There is no a test, because the tests work for the
localhost only anyway.

It still fails on Mac though in case the swim node was bound to
127.0.0.1. Then somewhy sendto() raises EADDRNOTAVAIL on attempt
to broadcast beyond the local machine. It happens on Linux too,
but with EINVAL error. These errors are ignored because are not
critical.

Part of #5864

(cherry picked from commit 1844eb6fcc3bb85b3240062e62af3321dbfef3fc)