sio: rework sio_strfaddr()
The function was overcomplicated, and made it harder to update it
in the next patches with functional changes.

The main source of the complication was usage of both inet_ntoa()
and getnameinfo(). The latter is more universal, it can cover the
case of the former.

The patch makes it use only getnameinfo() for IP addresses
regardless of v4 or v6.

Needed for #5632

(cherry picked from commit 441cb814d4cdd1ec4999b812d72e04cb5528687f)