box: simplify memory size options
Currently, you can specify the amount of memory for the configuration
only in bytes. It would also be convenient to be able to specify the
amount of memory in KB, MB or GB.

Add option to specify amount of memory as a string and a parser for this
string.

String must contain number of memory amount, suffix(KB, MB, or GB) and
any number of spaces at the begining, at the ending of the string or
between memory amount and suffix.

Closes #5786