box/lua: return exact nil from calls instead of empty return
Fix `luaT_pushtupleornil` and particular _ffi implementations
to return one nil argument instead of empty to change behaviour
of non-returning calls, like `:get`, `:update` or `:delete`
on non-existing keys.

Closes #3716

This patch changes behaviour of builtin functions.

The change is almost not seen from application point of view:
for almost any code there is no difference between
`return` and `return nil`, except for `select('#',...)`

But this change is visible in console and during yaml serialization:
Instead of empty yaml response
```
---
...
```
the response will be
```
---
- null
...
```

So, this affects many of test_run result files in such a way